requirements.txt
Flask==1.1.1
gunicorn==19.9.0
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
numpy==1.18.1
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
pandas==1.1.1
pynacl==1.4.0
